fix: update data types in ABI and models for liquidity intents
Changed lpTokenAmount from uint256 to uint128 in removeLiquidityIntentABI.

+++ src/abi.ts
@@ -120,7 +120,7 @@ export const removeLiquidityIntentABI = {
         },
         {
             name: "lpTokenAmount",
-            type: "uint256",
+            type: "uint128",
         },
         {
             name: "salt",
